requirements.txt
bcrypt>=4.0.1
passlib[bcrypt]==1.7.4
chromadb>=0.5.0
fastapi
uvicorn
pydantic
python-jose[cryptography]
pyjwt
httpx
PyYAML
pytest
pytest-asyncio
requests
aiosqlite
google-generativeai
python-dotenv
langchain-core
langchain-groq

langchain
langgraph>=0.0.20
numpy>=2.0.0
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
unstructured
reportlab
requests
python-multipart

# Use CPU-only torch to avoid huge CUDA downloads (~1.5GB savings)
--extra-index-url https://download.pytorch.org/whl/cpu
torch